# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KORION Chong API
  version: 0.1.0
paths:
...
  /api/merchant/dashboard:
    get:
      summary: Merchant dashboard page data
      parameters:
        - $ref: '#/components/parameters/MerchantIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          $ref: '#/components/responses/MerchantPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/merchant/transactions:
    get:
      summary: Merchant own payment and transaction page data
      parameters:
        - $ref: '#/components/parameters/MerchantIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
        - $ref: '#/components/parameters/PeriodQuery'
        - name: variant
          in: query
          schema:
            type: string
            enum:
              - all
              - refund
            default: all
      responses:
        '200':
          $ref: '#/components/responses/MerchantPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/merchant/payments:
    get:
      summary: Alias for merchant own payment and transaction page data
      parameters:
        - $ref: '#/components/parameters/MerchantIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          $ref: '#/components/responses/MerchantPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/merchant/store:
    get:
      summary: Merchant own store profile page data
      parameters:
        - $ref: '#/components/parameters/MerchantIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/MerchantPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/merchant/profile:
    get:
      summary: Merchant own account profile page data
      parameters:
        - $ref: '#/components/parameters/MerchantIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/MerchantPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/merchant/hq-notices:
    get:
      summary: Merchant received notices page data
      parameters:
        - $ref: '#/components/parameters/MerchantIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/MerchantPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/merchant/notices:
    get:
      summary: Alias for merchant received notices page data
      parameters:
        - $ref: '#/components/parameters/MerchantIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/MerchantPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/merchant/activity-logs:
    get:
      summary: Merchant own activity log page data
      parameters:
        - $ref: '#/components/parameters/MerchantIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/MerchantPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/merchant/settlements:
    get:
      summary: Merchant settlement history page data
      parameters:
        - $ref: '#/components/parameters/MerchantIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/MerchantPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/merchant/settlements/detail:
    get:
      summary: Merchant latest settlement detail page data
      parameters:
        - $ref: '#/components/parameters/MerchantIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/MerchantPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
components:
  parameters:
...
    MerchantIdHeader:
      name: X-Merchant-Id
      in: header
      required: true
      schema:
        type: integer
        format: int64
      description: Temporary merchant id until JWT auth is wired.
    CountryScopesHeader:
      name: X-Country-Scopes
      in: header
      required: true
      schema:
        type: string
      description: Temporary comma-separated country scopes until JWT auth is wired.
    CountryScopeQuery:
      name: countryScope
      in: query
      required: true
      schema:
        type: string
        minLength: 2
        maxLength: 2
      example: KR
    PeriodQuery:
      name: period
      in: query
      required: false
      schema:
        type: string
        pattern: '^\d{4}-\d{2}$'
      example: '2026-06'
...
  responses:
    ForbiddenCountryScope:
      description: Requested countryScope is outside the authenticated role's server-side country scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    MerchantPageData:
      description: Page data shaped for the current KORION PAY merchant admin frontend.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MerchantPageDataResponse'
  schemas:
...
    MerchantPageDataResponse:
      type: object
      additionalProperties: true
      description: >
        Transitional frontend page payload. Each merchant page returns the same top-level
        keys used by its existing JSON fallback, scoped by X-Merchant-Id and countryScope.
...
    ApiErrorResponse:
      type: object
      required:
        - timestamp
        - code
        - message
      properties:
        timestamp:
          type: string
          format: date-time
        code:
          type: string
        message:
          type: string
...